+++ src/WorldManagement/Configuration/WorldManagementConfigurationInstaller.cs
@@ -29,32 +29,16 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Sovereign.WorldManagement.Systems.WorldManagement
+namespace Sovereign.WorldManagement.Configuration
 {
-
-    /// <summary>
-    /// IoC installer for WorldManagementSystem support classes.
-    /// </summary>
-    public sealed class WorldManagementInstaller : IWindsorInstaller
+    public sealed class WorldManagementConfigurationInstaller : IWindsorInstaller
     {
         public void Install(IWindsorContainer container, IConfigurationStore store)
         {
-            container.Register(Component.For<WorldManagementEventHandler>()
-                .LifestyleSingleton());
-
             container.Register(EngineClasses.EngineAssemblies()
-                .BasedOn<IWorldSegmentLoader>()
+                .BasedOn<IWorldManagementConfiguration>()
                 .WithServiceDefaultInterfaces()
-                .LifestyleSingleton());
-
-            container.Register(EngineClasses.EngineAssemblies()
-                .BasedOn<IWorldSegmentUnloader>()
-                .WithServiceDefaultInterfaces()
-                .LifestyleSingleton());
-
-            container.Register(Component.For<WorldSegmentRegistry>()
                 .LifestyleSingleton());
         }
     }
-
 }